// wardn/frontend/lib/api/generated/model/invitationRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */
import type { InvitationReadScopeType } from './invitationReadScopeType';
import type { InvitationReadStatus } from './invitationReadStatus';

export interface InvitationRead {
  acceptedAt: string | null;
  acceptedById: string | null;
  createdAt: string;
  email: string;
  expiresAt: string;
  id: string;
  invitedById: string | null;
  organizationId: string;
  role: string;
  scopeType: InvitationReadScopeType;
  status: InvitationReadStatus;
  updatedAt: string;
  workspaceId: string | null;
}