// wardn/frontend/lib/api/generated/model/memberRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */
import type { MemberReadAccessSource } from './memberReadAccessSource';

export interface MemberRead {
  accessSource: MemberReadAccessSource;
  createdAt: string;
  displayName: string;
  email: string;
  firstName: string;
  lastName: string;
  membershipId: string | null;
  organizationRole?: string | null;
  role: string;
  userId: string;
}